save band, save set, mappers, service

+++ src/main/java/hr/tvz/java/zboroteka/repository/SongRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import hr.tvz.java.zboroteka.model.Song;
 
 @Repository
-public interface SongRepository extends JpaRepository<Song, Long> {
+public interface SongRepository extends JpaRepository<Song, Integer> {
 
 }
